GraduationDesign/requirements.txt
neo4j==5.15.0
python-dotenv==1.0.0
geopandas>=1.1.2